+++ src/styles/home.scss
@@ -1,15 +1,7 @@
 .hero {
   background-color: $primary;
   position: relative;
-
-  .tri-large {
-    opacity: .25;
-    position: absolute;
-    top: 3em;
-    left: 3em;
-    width: 75%;
-    z-index: 1;
-  }
+  height: 80vh;
   
   .content {
     align-items: flex-end;
@@ -43,6 +35,15 @@
   .pop {
     color: $grey1;
   }
+}
+
+.tri-large {
+  opacity: .25;
+  position: absolute;
+  top: 3em;
+  left: 3em;
+  width: 75%;
+  z-index: 1;
 }
 
 .tri-small {
