+++ src/styles/portfolio.scss
@@ -1,5 +1,9 @@
 
 .portfolio {
+  &.highlights {
+    margin-top: 3em;
+  }
+
   a {
     display: block;
     position: relative;
@@ -53,7 +57,7 @@
 
 .headline-image {
   width:100%;
-  height:100%;
+  height:80vh;
   overflow:hidden;
 
   img {
