+++ src/views/Portfolio.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 // import { Link } from "react-router-dom";
-import Projects from '../components/portfolioGallery';
+import Projects from '../components/projects';
 
 class Portfolio extends Component {
   render() {
